requirements.txt
watchdog>=4.0.0
fastmcp>=2.0
httpx>=0.27.0
obsidian-mcp-guard>=0.1.0
pyyaml>=6.0
pytest>=8.0
pytest-cov>=5.0
